Convert product parent id from integer into long

+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/persistence/entity/ProductEntity.kt
@@ -66,7 +66,7 @@ data class ProductEntity(
     val averageRating: String? = null,
     val ratingCount: Int? = null,
 
-    val parentId: Int? = null,
+    val parentId: Long? = null,
     val purchaseNote: String? = null,
     val menuOrder: Int? = null
 )
